# src/dscompanion/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dscompanion API",
    "description": "REST surface over dscompanion's 13-step interactive pipeline wizard.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs/{run_id}/steps/train/preview": {
      "post": {
        "tags": [
          "train"
        ],
        "summary": "Preview",
        "description": "Reports the available algorithms and feature-column count.\n\nArgs:\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    TrainPreviewResponse: Algorithm picker options and feature count.\n\nRaises:\n    HTTPException: 400, if Steps 1-3 have not been confirmed for this run.",
        "operationId": "preview_api_runs__run_id__steps_train_preview_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/train/leaderboard/start": {
      "post": {
        "tags": [
          "train"
        ],
        "summary": "Leaderboard Start",
        "description": "Starts a background leaderboard comparison across every supported algorithm.\n\nArgs:\n    run (RunState): Injected via ``get_pipeline_service``.\n    store (RunStateStore): Injected via ``get_state_store``.\n\nReturns:\n    LeaderboardStartResponse: The new job's id \u2014 poll\n    ``GET /api/runs/{run_id}/jobs/{job_id}`` for progress and the result.\n\nRaises:\n    HTTPException: 400, if Steps 1-3 have not been confirmed, or no\n        feature columns remain.",
        "operationId": "leaderboard_start_api_runs__run_id__steps_train_leaderboard_start_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/train/confirm": {
      "post": {
        "tags": [
          "train"
        ],
        "summary": "Confirm",
        "description": "Trains (single-algorithm path) or accepts an already-trained leaderboard\nwinner, then persists the result.\n\nArgs:\n    body (TrainConfirmRequest): The algorithm to confirm.\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    TrainConfirmResponse: Confirmation and a brief metrics summary.\n\nRaises:\n    HTTPException: 400, if Steps 1-3 have not been confirmed, or training\n        fails.",
        "operationId": "confirm_api_runs__run_id__steps_train_confirm_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/tuning/preview": {
      "post": {
        "tags": [
          "tuning"
        ],
        "summary": "Preview",
        "description": "Reports whether tuning is available for Step 8's confirmed algorithm.\n\nArgs:\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    TuningPreviewResponse: Availability and slider bounds.\n\nRaises:\n    HTTPException: 400, if Step 8 has not been confirmed for this run.",
        "operationId": "preview_api_runs__run_id__steps_tuning_preview_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TuningPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/tuning/tune/start": {
      "post": {
        "tags": [
          "tuning"
        ],
        "summary": "Tune Start",
        "description": "Starts a background Optuna tuning run.\n\nArgs:\n    body (TuningStartRequest): Number of trials to run.\n    run (RunState): Injected via ``get_pipeline_service``.\n    store (RunStateStore): Injected via ``get_state_store``.\n\nReturns:\n    TuningStartResponse: The new job's id \u2014 poll\n    ``GET /api/runs/{run_id}/jobs/{job_id}`` for progress and the result.\n\nRaises:\n    HTTPException: 400, if Step 8 has not been confirmed for this run.",
        "operationId": "tune_start_api_runs__run_id__steps_tuning_tune_start_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TuningStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TuningStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/tuning/confirm": {
      "post": {
        "tags": [
          "tuning"
        ],
        "summary": "Confirm",
        "description": "Persists the final model choice (tuned or Step 8's default).\n\nArgs:\n    body (TuningConfirmRequest): Whether to accept the tuned model.\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    TuningConfirmResponse: Confirmation, whether tuned, and the best score.\n\nRaises:\n    HTTPException: 400, if Step 8 has not been confirmed, or\n        ``use_tuned=True`` but no tuned model is available.",
        "operationId": "confirm_api_runs__run_id__steps_tuning_confirm_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TuningConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TuningConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeaderboardStartResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "LeaderboardStartResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/train/leaderboard/start``.\n\nArgs:\n    job_id (str): Poll ``GET /api/runs/{run_id}/jobs/{job_id}`` with this\n        id for live per-algorithm progress and the final ranked table\n        (``result`` shape: ``{\"rows\": [{\"algorithm\", \"status\", \"fit_time_s\",\n        \"roc_auc\", \"error\"}, ...], \"winner\": str | None}``)."
      },
...
      "TrainAlgorithmOption": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "name",
          "note"
        ],
        "title": "TrainAlgorithmOption",
        "description": "One selectable algorithm for Step 8's picker.\n\nArgs:\n    name (str): Algorithm key (e.g. ``\"xgboost\"``), as accepted by\n        ``ModelFactory.build()``.\n    note (str): Plain-language one-liner shown next to it."
      },
      "TrainConfirmRequest": {
        "properties": {
          "algorithm": {
            "type": "string",
            "title": "Algorithm"
          }
        },
        "type": "object",
        "required": [
          "algorithm"
        ],
        "title": "TrainConfirmRequest",
        "description": "Request for ``POST /api/runs/{run_id}/steps/train/confirm``.\n\nArgs:\n    algorithm (str): The algorithm to train and confirm (single-algorithm\n        path), or the leaderboard winner's name (leaderboard path \u2014 the\n        already-trained winner is read from ``run.artifacts`` rather than\n        retrained)."
      },
      "TrainConfirmResponse": {
        "properties": {
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          },
          "algorithm": {
            "type": "string",
            "title": "Algorithm"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/TrainMetricRow"
            },
            "type": "array",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "confirmed",
          "algorithm",
          "metrics"
        ],
        "title": "TrainConfirmResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/train/confirm``.\n\nArgs:\n    confirmed (bool): Always ``True`` on success.\n    algorithm (str): The confirmed algorithm.\n    metrics (list[TrainMetricRow]): Brief per-split metric summary."
      },
      "TrainMetricRow": {
        "properties": {
          "split": {
            "type": "string",
            "title": "Split"
          },
          "auc_roc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auc Roc"
          },
          "gini": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gini"
          },
          "ks_statistic": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ks Statistic"
          }
        },
        "type": "object",
        "required": [
          "split"
        ],
        "title": "TrainMetricRow",
        "description": "One row of Step 8's brief AUC/Gini/KS summary table.\n\nArgs:\n    split (str): One of ``\"train\"``, ``\"val\"``, ``\"test\"``, ``\"oot\"``.\n    auc_roc (float | None): AUC-ROC, ``None`` if unavailable for this split.\n    gini (float | None): Gini coefficient, ``None`` if unavailable.\n    ks_statistic (float | None): KS statistic, ``None`` if unavailable."
      },
      "TrainPreviewResponse": {
        "properties": {
          "algorithms": {
            "items": {
              "$ref": "#/components/schemas/TrainAlgorithmOption"
            },
            "type": "array",
            "title": "Algorithms"
          },
          "leaderboard_algorithms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Leaderboard Algorithms"
          },
          "n_features": {
            "type": "integer",
            "title": "N Features"
          }
        },
        "type": "object",
        "required": [
          "algorithms",
          "leaderboard_algorithms",
          "n_features"
        ],
        "title": "TrainPreviewResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/train/preview``.\n\nArgs:\n    algorithms (list[TrainAlgorithmOption]): Every algorithm available for\n        the single-algorithm path (task-appropriate).\n    leaderboard_algorithms (list[str]): Subset of ``algorithms`` used for\n        the leaderboard comparison \u2014 excludes ``\"svm\"`` (too slow for an\n        automated multi-algorithm run).\n    n_features (int): Feature-column count after Steps 2/3/5/6 exclusions,\n        before any transform runs."
      },
...
      "TuningConfirmRequest": {
        "properties": {
          "use_tuned": {
            "type": "boolean",
            "title": "Use Tuned"
          }
        },
        "type": "object",
        "required": [
          "use_tuned"
        ],
        "title": "TuningConfirmRequest",
        "description": "Request for ``POST /api/runs/{run_id}/steps/tuning/confirm``.\n\nArgs:\n    use_tuned (bool): ``True`` to carry the tuned model forward, ``False``\n        to keep Step 8's model unchanged \u2014 covers the skip-toggle,\n        auto-skip, and reject-tuned-parameters paths alike, all three\n        confirm with ``use_tuned=False``."
      },
      "TuningConfirmResponse": {
        "properties": {
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          },
          "tuned": {
            "type": "boolean",
            "title": "Tuned"
          },
          "best_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Score"
          }
        },
        "type": "object",
        "required": [
          "confirmed",
          "tuned"
        ],
        "title": "TuningConfirmResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/tuning/confirm``.\n\nArgs:\n    confirmed (bool): Always ``True`` on success.\n    tuned (bool): Whether the tuned model was accepted.\n    best_score (float | None): Best AUC found by tuning, if applicable."
      },
      "TuningPreviewResponse": {
        "properties": {
          "algorithm": {
            "type": "string",
            "title": "Algorithm"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "default_n_trials": {
            "type": "integer",
            "title": "Default N Trials"
          },
          "min_n_trials": {
            "type": "integer",
            "title": "Min N Trials"
          },
          "max_n_trials": {
            "type": "integer",
            "title": "Max N Trials"
          }
        },
        "type": "object",
        "required": [
          "algorithm",
          "available",
          "default_n_trials",
          "min_n_trials",
          "max_n_trials"
        ],
        "title": "TuningPreviewResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/tuning/preview``.\n\nArgs:\n    algorithm (str): The algorithm confirmed at Step 8.\n    available (bool): ``False`` when ``algorithm`` has no Optuna search\n        space (``gradient_boosting``/``lightgbm``) \u2014 the frontend should\n        auto-skip straight to confirming with ``use_tuned=False``,\n        matching Streamlit's silent auto-skip rather than offering a\n        \"Start Tuning\" control that would only fail.\n    default_n_trials (int): Suggested starting value for the trials slider.\n    min_n_trials (int): Slider minimum.\n    max_n_trials (int): Slider maximum."
      },
      "TuningStartRequest": {
        "properties": {
          "n_trials": {
            "type": "integer",
            "title": "N Trials"
          }
        },
        "type": "object",
        "required": [
          "n_trials"
        ],
        "title": "TuningStartRequest",
        "description": "Request for ``POST /api/runs/{run_id}/steps/tuning/tune/start``.\n\nArgs:\n    n_trials (int): Number of Optuna trials to run."
      },
      "TuningStartResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "TuningStartResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/tuning/tune/start``.\n\nArgs:\n    job_id (str): Poll ``GET /api/runs/{run_id}/jobs/{job_id}`` with this\n        id \u2014 ``result`` holds the full comparison payload once done:\n        ``{\"best_score\": float | None, \"baseline_auc\": float | None,\n        \"params_comparison\": [{\"parameter\", \"default\", \"tuned\", \"changed\"},\n        ...], \"trials\": [{\"trial_number\", \"metric_value\"}, ...],\n        \"optimization_curve\": <Plotly figure JSON | None>}``. ``baseline_auc``\n        is Step 8's untuned model re-evaluated fresh (val split, falling back\n        to test), for the before/after comparison."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}